python_scripts/requirements.txt
langchain== 0.2.10
langchain-openai==0.1.1
pydantic==2.5.1
python-dotenv==1.0.0
jinja2==3.1.2
javalang==0.13.0
huggingface-hub==0.19.4
transformers==4.36.0
tokenizers==0.15.0
openai=1.3.5